+++ src/netdiffuser_extra.cpp
@@ -19,24 +19,8 @@ arma::umat sparse_indexes(const arma::sp_mat & mat) {
   int i = 0;
   for (arma::sp_mat::const_iterator it = begin; it != end; ++it) {
     indices.at(i, 0) = it.row();
-    indices.at(i, 1) = it.col();
-
-    // Incrementing
-    i++;
+    indices.at(i++, 1) = it.col();
   }
-
-  // int curcol = 1;
-  // for (int i=0;i<n;i++) {
-  //
-  //   // Incrementing column while there's no new element in the col
-  //   while (mat.col_ptrs[curcol] <= i) ++curcol;
-  //   // while (mat.col_ptrs[curcol-1] == mat.col_ptrs[curcol]) ++curcol;
-  //
-  //   // Asigning indexes
-  //   indices.at(i, 0) = mat.row_indices[i];
-  //   // indices.at(i,1) = j;
-  //   indices.at(i, 1) = curcol - 1;
-  // }
 
   // return indices;
   return indices;
@@ -77,13 +61,6 @@ arma::sp_mat sp_trimatl(const arma::sp_mat & x) {
     if (i >= j)
       newx.at(i,j) = *it;
   }
-
-  // arma::umat indices = sparse_indexes(x);
-  // for (int i=0; i< (int) indices.n_rows; i++) {
-  //   if (indices.at(0,i) >= indices.at(1,i))
-  //     newx.at(indices.at(0,i), indices.at(1,i)) =
-  //       x.at(indices.at(0,i), indices.at(1,i));
-  // }
 
   return newx;
 }